Start incorporating the necessary code
